Clean up Header and Footer and move files
