+++ src/components/Footer.js
@@ -1,0 +1,3 @@
+export default function Footer() {
+  return <></>;
+}
