--- src/header-footer/Footer.js
@@ -1,5 +1,0 @@
-function Footer() {
-  return <></>;
-}
-
-export default Footer;
